python/kafka-cons/requirements.txt
Flask==1.1.1
werkzeug==0.16.1
requests
kafka-python